handle the no-releases-to-cleanup case; drive-by shellcheck/formatting fixes

+++ hack/bin/integration-cleanup.sh
@@ -6,14 +6,18 @@
 
 set -euo pipefail
 
-DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
-TOP_LEVEL="$DIR/../.."
+DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 
-releases=$(helm list -A -f 'test-' -o json \
-               | jq -r -f "$DIR/filter-old-releases.jq")
-while read line; do
-    name=$(awk '{print $1}' <<<"$line")
-    namespace=$(awk '{print $2}' <<<"$line")
-    echo "test release '$name' older than 2 hours; deleting..."
-    helm delete -n "$namespace" "$name"
-done <<<"$releases"
+releases=$(helm list -A -f 'test-' -o json |
+    jq -r -f "$DIR/filter-old-releases.jq")
+
+if [ -n "$releases" ]; then
+    while read -r line; do
+        name=$(awk '{print $1}' <<<"$line")
+        namespace=$(awk '{print $2}' <<<"$line")
+        echo "test release '$name' older than 2 hours; deleting..."
+        helm delete -n "$namespace" "$name"
+    done <<<"$releases"
+else
+    echo "Nothing to clean up."
+fi
